\name{pecCforest}
\alias{pecCforest}
\title{
S3-wrapper function for cforest from the party package
}
\description{
S3-wrapper function for cforest from the party package
}
\usage{
pecCforest(formula, data, ...)
}
\arguments{
  \item{formula}{
    See \code{cforest} of the \code{party} package
  }
  \item{data}{
    See \code{cforest} of the \code{party} package
}
  \item{\dots}{
    See \code{cforest} of the \code{party} package
}
}
\details{
    See \code{cforest} of the \code{party} package.
}
\references{
Mogensen, UB and Ishwaran, H and Gerds, TA (2010)
Evaluating random forests for survival analysis using prediction error curves
Research report 10/8. Department of Biostatistics, University of
Copenhagen
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{survival}
