% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predictSurvProb.ctree.R
\name{pecCtree}
\alias{pecCtree}
\title{S3-Wrapper for ctree.}
\usage{
pecCtree(...)
}
\arguments{
\item{...}{passed to ctree}
}
\value{
list with two elements: ctree and call
}
\description{
The call is added to an ctree object
}
\examples{
library(prodlim)
library(party)
library(survival)
set.seed(50)
d <- SimSurv(50)
nd <- data.frame(X1=c(0,1,0),X2=c(-1,0,1))
f <- pecCtree(Surv(time,status)~X1+X2,data=d)
predictSurvProb(f,newdata=nd,times=c(3,8))
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
\seealso{
pecCforest
}

