% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.calibrationPlot.R
\name{plot.calibrationPlot}
\alias{plot.calibrationPlot}
\title{Plot objects obtained with \code{calPlot}}
\usage{
\method{plot}{calibrationPlot}(x, ...)
}
\arguments{
\item{x}{Object obtained with \code{calPlot}}

\item{...}{Not used.}
}
\value{
Nothing
}
\description{
Calibration plots
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
\seealso{
\code{calPlot}
}

