% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.riskReclassification.R
\name{print.riskReclassification}
\alias{print.riskReclassification}
\title{Print riskReclassification object}
\usage{
\method{print}{riskReclassification}(x, percent = TRUE, digits = ifelse(percent, 1, 2), ...)
}
\arguments{
\item{x}{Object obtained with function \link{reclass}}

\item{percent}{Logical. If \code{TRUE} multiply risks with 100.}

\item{digits}{Rounding digits}

\item{...}{Not used}
}
\description{
Print method for results of reclass
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
