\name{simCost}
\alias{simCost}
\title{Simulate COST alike data}
\usage{
simCost(N)
}
\arguments{
  \item{N}{Sample size}
}
\value{
Data frame
}
\description{
Simulate data alike the data from the Copenhagen stroke
study (COST)
}
\details{
This uses functionality of the lava package.
}
\author{
Thomas Alexander Gerds
}

