% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pec-package.R
\docType{data}
\name{threecity}
\alias{threecity}
\title{threecity data}
\format{A subsample consisting of 2000 observations on the following 3 variables.
\describe{ \item{pi}{5-year absolute risk predictions of dementia.}
\item{status}{0=censored, 1=dementia, 2=death dementia free}
\item{time}{time to event (i.e., time to
either dementia, death dementia free or loss of follow-up)}}}
\source{
Web-appendix of Blanche et al. (2015).
}
\description{
Extracted data from a french population based cohort (Three-City cohort). The dataset includes 
followup information on dementia outcome and predicted 5-year risks based on 
based on the subject specific information which includes age, gender,
education level and cognitive decline measured by a psychometric test
(Mini Mental State Examination). The prediction model from which the
predictions have been computed has been fitted on independent training
data from the Paquid cohort, another french population based cohort with similar design (see Reference Blanche et al. 2015 for details) .
}
\examples{

data(threecity)

}
\references{
Blanche, P., Proust-Lima, C., Loubere, L., Berr, C., Dartigues, J. F., Jacqmin-Gadda, H. (2015).
Quantifying and comparing dynamic
predictive accuracy of joint models for longitudinal marker and
time-to-event in presence of censoring and competing risks. 
Biometrics, 71(1), 102-113.
}
\keyword{datasets}

