% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pec-package.R
\docType{data}
\name{Pbc3}
\alias{Pbc3}
\title{Pbc3 data}
\format{A data frame with 349 observations on the following 15 variables.
\describe{ \item{ptno}{patient identification}
\item{unit}{hospital (1: Hvidovre, 2: London, 3: Copenhagen, 4:
Barcelona, 5: Munich, 6: Lyon)} \item{tment}{treatment (0: placebo,
1: CyA)} \item{sex}{(1: males, 0: females)} \item{age}{age
in years} \item{stage}{histological stage (1, 2, 3, 4)}
\item{gibleed}{previous gastrointestinal bleeding (1: yes, 0: no)}
\item{crea}{creatinine (micromoles/L)} \item{alb}{albumin
(g/L)} \item{bili}{bilirubin (micromoles/L)}
\item{alkph}{alkaline phosphatase (IU/L)}
\item{asptr}{aspartate transaminase (IU/L)}
\item{weight}{body weight (kg)} \item{days}{observation time
(days)} \item{status}{status at observation time (0: censored, 1:
liver transplantation, 2 : dead)} }}
\source{
\url{http://192.38.117.59/~linearpredictors/?page=datasets&dataset=Pbc3}
}
\description{
PBC3 was a multi-centre randomized clinical trial conducted in six European
hospitals. Between 1 Jan. 1983 and 1 Jan. 1987, 349 patients with the liver
disease primary biliary cirrhosis (PBC) were randomized to either treatment
with Cyclosporin A (CyA, 176 patients) or placebo (173 patients). The
purpose of the trial was to study the effect of treatment on the survival
time. However, during the course of the trial an increased use of liver
transplantation for patients with this disease made the investigators
redefine the main response variable to be time to ``failure of medical
treatment'' defined as either death or liver transplantation. Patients were
then followed from randomization until treatment failure, drop-out or 1 Jan,
1989; 61 patients died (CyA: 30, placebo: 31), another 29 were transplanted
(CyA: 14, placebo: 15) and 4 patients were lost to follow-up before 1 Jan.
1989. At entry a number of clinical, biochemical and histological variables,
including serum bilirubin, serum albumin, sex, age were recorded.
}
\examples{
data(Pbc3)
}
\references{
Andersen and Skovgaard. Regression with linear predictors.
Springer, 2010.
}
\keyword{datasets}

