\name{pecCforest}
\alias{pecCforest}
\title{
S3-wrapper function for cforest from the party package
}
\description{
S3-wrapper function for cforest from the party package
}
\usage{
pecCforest(formula, data, ...)
}
\arguments{
  \item{formula}{
    See \code{cforest} of the \code{party} package
  }
  \item{data}{
    See \code{cforest} of the \code{party} package
}
  \item{\dots}{
    See \code{cforest} of the \code{party} package
}
}
\details{
    See \code{cforest} of the \code{party} package.
}
\references{
  Ulla B. Mogensen, Hemant Ishwaran, Thomas A. Gerds (2012). Evaluating Random Forests
  for Survival Analysis Using Prediction Error Curves. Journal of Statistical Software,
  50(11), 1-23. URL http://www.jstatsoft.org/v50/i11/.
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{survival}
