% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reclass.R
\name{reclass}
\alias{reclass}
\title{Risk reclassification table}
\usage{
reclass(list, formula, newdata, time, cause, cuts = seq(0, 100, 25),
  digits = 1)
}
\arguments{
\item{list}{A list with two elements. Each element should either
be a vector with probabilities, or an object for which \code{predictStatusProb}
can extract predicted risk based on newdata.}

\item{newdata}{Passed on to \code{predictStatusProb}}

\item{cuts}{Risk quantiles to group risk}

\item{digits}{Number of digits to show for the predicted risks}
}
\value{
reclassification table
}
\description{
Tabulate grouped risks predicted by two different methods, models, algorithms
}
\details{
All risks are multiplied by 100 before
}
\examples{
library(survival)
set.seed(40)
d <- prodlim::SimSurv(140)
nd <- prodlim::SimSurv(140)
Models <- list("Cox.X2"=coxph(Surv(time,status)~X2,data=d),
               "Cox.X1.X2"=coxph(Surv(time,status)~X1+X2,data=d))
rc <- reclass(Models,Surv(time,status)~1,newdata=nd,time=5)
print(rc)
plot(rc)
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
\seealso{
predictStatusProb
}

