% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wallyPlot.R
\name{wallyPlot}
\alias{wallyPlot}
\title{Wally plots to assess calibration of a risk or survival prediction}
\usage{
wallyPlot(object, time, formula, data, cause = 1, q = 10, ylim,
  hanging = FALSE, seed = NULL, mar = c(4.1, 4.1, 2, 2), colbox = "red",
  verbose = TRUE, ...)
}
\arguments{
\item{object}{Risk or survival prediction. Either a numeric vector
of predictions or an object which in the simple survival setting
has a \code{predictSurvProb} method and in the competing risks
setting has a \code{predictEventProb} method.}

\item{time}{Time interest for evaluating calibration of the
predictions.}

\item{formula}{A survival or event history formula. The left hand
side is used to compute the expected event status. If
\code{formula} is \code{missing}, try to extract a formula from
the first element in object.}

\item{data}{A data frame in which to validate the prediction
models and to fit the censoring model. If \code{data} is missing,
try to extract a data set from the first element in object.}

\item{cause}{For competing risks settings the cause of interest.}

\item{q}{The number of quantiles. Defaults to 10.}

\item{ylim}{Limits of y-axis. If missing the function tries to
find appropriate limits based on the simulated and real data.}

\item{hanging}{If \code{TRUE}, hang bars corresponding to observed
frequencies at the value of the corresponding prediction.}

\item{seed}{A seed value to make results reproducible.}

\item{mar}{Plot margins passed to par.}

\item{colbox}{Color of the box which identifies the real data
calibration plot.}

\item{verbose}{If \code{TRUE} warn about missing formula and data.}

\item{...}{Further arguments passed to \code{calPlot}.}
}
\value{
List of simulated and real data.
}
\description{
Wally plots to assess calibration of a risk or survival prediction
}
\examples{
# Survival setting
library(prodlim)
library(survival)
d = SimSurv(180)
f = coxph(Surv(time,status)~X1+X2,data=d)
\dontrun{wallyPlot(f,time=2,q=4,data=d,formula=Surv(time,status)~1)}
}
\author{
Paul F. Blanche <paul.blanche@univ-ubs.fr> and Thomas A. Gerds <tag@biostat.ku.dk>
}
\seealso{
calPlot
}

